fix(Upload): if no file found in the folder provided to upload the file, provide a console message, and not traceback
Earlier, while uploading the files to a dataset, if the user has provided a blank folder in the path, they get a traceback, making an appearance like something had failed. Also, an import id was being generated by the API. Now no import ID is being generated either, and user is being provided an explicit message that there were no DICOM files in the provided folder.